SQL
Modificacion en las tablas de bdd

+++ ms-tarifas/src/main/java/cl/hilton/tarifas/model/ProjTipoHabitacion.java
@@ -23,7 +23,7 @@ public class ProjTipoHabitacion {
     private String descripcion;
 
     @Column(name = "capacidad_max", nullable = false)
-    private Short capacidadMax;
+    private long capacidadMax;
 
     @Column(name = "actualizado_en", nullable = false)
     private OffsetDateTime actualizadoEn;
